Hw 04 auth

+++ src/services/contacts.js
@@ -8,38 +8,52 @@ class ContactsService {
     })
   }
 
-  async listContacts() {
-    const contacts = await this.model.find({})
-    console.log(contacts)
-    return contacts
+  async listContacts(userId, { limit = 20, offset = 0, sortBy, sortByDesc, filter }) {
+    const { docs: contacts, totalDocs: total } = await this.model.paginate(
+      { owner: userId },
+      {
+        limit,
+        offset,
+        sort: {
+          ...sortBy ? { [`${sortBy}`]: 1 } : {},
+          ...sortByDesc ? { [`${sortByDesc}`]: -1 } : {}
+        },
+        select: filter ? filter.split('|').join('') : '',
+        populate: {
+          path: 'owner',
+          select: 'name email phone'
+        }
+      }
+    )
+    return { contacts, total, limit: Number(limit), offset: Number(offset) }
   }
 
-  async getContactById(contactId) {
-    // const objectId = ObjectID(contactId)
-    const contactById = await this.model.findOne({ _id: contactId })
+  async getContactById(userId, contactId) {
+    const contactById = await this.model.findOne({ _id: contactId, owner: userId }).populate({
+      path: 'owner',
+      select: 'name email phone'
+    })
     console.log(contactById)
     return contactById
   }
 
-  async addContact(name, email, phone) {
-    const contact = await this.model.create({ name, email, phone })
+  async addContact(userId, body) {
+    const contact = await this.model.create({ ...body, owner: userId })
     console.log(contact)
     return contact
   }
 
-  async updateContact(contactId, body) {
-    // const objectId = ObjectID(contactId)
+  async updateContact(userId, contactId, body) {
     const contact = await this.model.findByIdAndUpdate(
-      { _id: contactId },
+      { _id: contactId, owner: userId },
       { ...body },
       { new: true },
     )
     return contact
   }
 
-  async removeContact(contactId) {
-    // const objectId = ObjectID(contactId)
-    const contact = await this.model.findByIdAndRemove({ _id: contactId })
+  async removeContact(userId, contactId) {
+    const contact = await this.model.findByIdAndRemove({ _id: contactId, owner: userId })
     console.log(contact)
     return contact
   }
